+++ mrb_doc/core/drawing.rb
@@ -40,6 +40,23 @@ def end_mode2D()
   nil
 end
 
+# Starts drawing to the RenderTexture
+# @param texture [RenderTexture]
+# @return [nil]
+def begin_texture_mode(texture)
+  # mrb_begin_texture_mode
+  # src/mruby_integration/core/drawing.cpp
+  nil
+end
+
+# Stops drawing to the RenderTexture
+# @return [nil]
+def end_texture_mode()
+  # mrb_end_texture_mode
+  # src/mruby_integration/core/drawing.cpp
+  nil
+end
+
 # In scissor mode only draw calls within the defined area will actually be drawn
 # @param x [Integer]
 # @param y [Integer]